sentiment_pipeline/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
yfinance==0.2.44
transformers==4.45.2
torch==2.4.1
datasets==3.0.1
peft==0.13.1
bitsandbytes==0.43.2
tqdm==4.66.5
matplotlib==3.9.2
python-dotenv==1.0.1
pyyaml==6.0.2
schedule==1.2.1
accelerate==0.34.2
wandb==0.18.3
trl==0.9.6
feedparser==6.0.11

